# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Socious-API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://dev.socious.io/api/v2/"
    },
    {
      "url": "http://localhost:8370"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/projects": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "get all",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "project"
        ],
        "summary": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "description": "Data testing",
                  "experience_level": 1,
                  "payment_range_higher": "13",
                  "payment_range_lower": "12",
                  "payment_scheme": "FIXED",
                  "payment_type": "VOLUNTEER",
                  "remote_preference": "HYBRID",
                  "social_causes": [
                    "SOCIAL",
                    "POVERTY"
                  ],
                  "status": "ACTIVE",
                  "title": "Testing"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/projects/update/{project_id}": {
      "post": {
        "tags": [
          "project"
        ],
        "summary": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "causes_tags": [
                    "SOCIAL"
                  ],
                  "country": "AL",
                  "created_at": "2022-09-05T17:19:25.501Z",
                  "deleted_at": null,
                  "description": "Structure dsdsd",
                  "experience_level": 1,
                  "expires_at": null,
                  "id": "af8f8687-3cc9-49db-8231-4636cfb9a3a5",
                  "identity_id": "ebd02a72-f530-4d18-9a72-d7aa9ace1271",
                  "old_id": null,
                  "passions": [
                    "SOCIAL"
                  ],
                  "payment_currency": null,
                  "payment_range_higher": "13",
                  "payment_range_lower": "12",
                  "payment_scheme": "HOURLY",
                  "payment_type": "PAID",
                  "project_length": "LESS_THAN_A_DAY",
                  "project_type": "PART_TIME",
                  "remote_preference": "HYBRID",
                  "skills": [
                    "ANGULARJS",
                    "CSS",
                    "AWS"
                  ],
                  "status": "ACTIVE",
                  "title": "Sdsdsd Demo New ",
                  "updated_at": "2022-09-05T17:19:25.501Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "Current-Identity",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "ebd02a72-f530-4d18-9a72-d7aa9ace1271"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/projects/{project_id}/questions": {
      "post": {
        "tags": [
          "project"
        ],
        "summary": "add question",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "question": "are you OK ?",
                  "required": false,
                  "options": [
                    "yes",
                    "no"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/projects/{project_id}/applicants": {
      "parameters": [
        {
          "name": "project_id",
          "in": "path",
          "schema": {
            "type": "string"
          },
          "required": true
        }
      ],
      "post": {
        "tags": [
          "project"
        ],
        "summary": "apply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "cover_letter": "test text",
                  "cv_link": "http://test.com/cv.pdf",
                  "cv_name": "test text",
                  "share_contact_info": true,
                  "answers": [
                    {
                      "id": "3fcc4216-0d0a-412d-9747-494e4dc8e53f",
                      "answer": "yes"
                    },
                    {
                      "id": "7b9df588-50b2-4896-b343-b9f99c9b7ee1",
                      "selected_option": 1,
                      "answer": "yes"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/user/profile": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Current-Identity",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "311fca4c-057d-4752-9e7c-b985ff1c5294"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/user/update/profile": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "update profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "bio": "I’m looking right now through the code. typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more re dfsdfgdfg Dfg",
                  "city": "San Pablo City",
                  "country": "PH",
                  "first_name": "Jayson",
                  "last_name": "Ilagan",
                  "mission": "Hdhshsh",
                  "mobile_country_code": "+63",
                  "phone": "5464542",
                  "skills": [
                    "ANGULARJS",
                    "JAVASCRIPT",
                    "NODE.JS",
                    "REACT.JS",
                    "REACT_NATIVE"
                  ],
                  "social_causes": [
                    "POVERTY",
                    "HEALTH",
                    "CHILD_PROTECTION",
                    "ANIMAL_RIGHTS",
                    "DEMOCRACY"
                  ],
                  "username": "socious2022"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/user/change-password": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "change password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "current_password": "socious2022",
                  "password": "socious2023"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/user/change-password-direct": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "change password direct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "password": "socious2022"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/user/delete": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "reason": "test"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/user/{user_id}/profile": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "schema": {
            "type": "string"
          },
          "required": true
        }
      ],
      "get": {
        "tags": [
          "users"
        ],
        "summary": "other user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/user/by-username/{username}/profile": {
      "parameters": [
        {
          "name": "username",
          "in": "path",
          "schema": {
            "type": "string"
          },
          "example": "seira-yun8611",
          "required": true
        }
      ],
      "get": {
        "tags": [
          "users"
        ],
        "summary": "other profile by username",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Current-Identity",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "311fca4c-057d-4752-9e7c-b985ff1c5294"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "q": "search q",
                "type": "projects"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/search/history": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "current user search history",
        "parameters": [
          {
            "name": "Current-Identity",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "c8765d15-030d-4ebc-a493-9265c0911f16"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "example": 100
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "example": 1
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}